Use sealed interface.
